implementations/pyspark/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kimball_framework"
version = "0.1.1"
description = "A declarative, CDF-based ETL framework for Databricks"
authors = [{name = "Tiago Marques"}]
license = {text = "Apache-2.0"}
dependencies = [
    "pyyaml",
    "jinja2",
    "jsonschema>=4.0.0",
    "pydantic>=2.0.0"
]


[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy"
]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["examples"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["examples"]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportMissingImports = "none"
exclude = ["examples"]

[tool.vulture]
min_confidence = 80
paths = ["src"]
exclude = ["examples"]
sort_by_size = true
verbose = false
